+++ src/util.js
@@ -1,5 +1,3 @@
-import {RenderPosition} from './constants.js';
-
 export const getRandomInt = (min, max) => min + Math.round(max * Math.random());
 export const getElement = (dom, identity) => dom.querySelector(identity);
 
@@ -28,20 +26,4 @@ export const createElement = (template) => {
   newElement.innerHTML = template;
 
   return newElement.firstChild;
-};
-
-export const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  if (container && element) {
-    switch (place) {
-      case RenderPosition.AFTERBEGIN:
-        container.prepend(element);
-        break;
-      case RenderPosition.BEFOREEND:
-        container.append(element);
-        break;
-      case RenderPosition.AFTEREND:
-        container.after(element);
-        break;
-    }
-  }
 };
